+++ phy/cluster/views/correlogram.py
@@ -42,6 +42,7 @@ class CorrelogramView(ScalingMixin, ManualClusteringView):
 
     # Do not show too many clusters.
     max_n_clusters = 20
+    defer_hidden_updates = True
 
     _default_position = 'left'
     cluster_ids = ()